Allow to use with Sprockets without Rails
