+++ lib/csso-rails.rb
@@ -1,7 +1,2 @@
 require 'csso'
-
-if defined?(Rails)
-  require 'csso/rails'
-else
-  require 'csso'
-end
+require 'csso/rails' if defined?(Rails)
