+++ lib/csso/rails.rb
@@ -6,7 +6,7 @@ module Csso
 
   class Railtie < ::Rails::Railtie
     initializer "csso.environment", :after => "sprockets.environment" do
-      CssCompressor.register
+      CssCompressor.register!
     end
 
     # saas-rails-3.2.4(and may be others) sets itself as default, ignoring config? => override :(
@@ -16,22 +16,6 @@ module Csso
       end
     end
 
-  end
-
-  class CssCompressor
-    def compress(css)
-      require 'csso'
-      #TODO: settings?
-      Csso.optimize(css, true)
-    end
-
-    def self.register
-      if Sprockets.respond_to? :register_compressor
-        Sprockets.register_compressor('text/css', COMPRESSOR_SYM, 'Csso::CssCompressor')
-      else
-        Sprockets::Compressors.register_css_compressor(COMPRESSOR_SYM, 'Csso::CssCompressor', :default => true)
-      end
-    end
   end
 
 end
